删除角色、菜单时校验

+++ api/src/admin/system/role/role.service.ts
@@ -7,12 +7,15 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { RoleEntity } from './role.entity';
 import { In, Like, Repository } from 'typeorm';
 import { RolePageListDto } from './role.dto';
+import { UserEntity } from '../user/user.entity';
 
 @Injectable()
 export class RoleService {
   constructor(
     @InjectRepository(RoleEntity)
     private readonly roleRepository: Repository<RoleEntity>,
+    @InjectRepository(UserEntity)
+    private readonly userRepository: Repository<UserEntity>,
   ) {}
 
   async getRoleList(menuPageListDto: RolePageListDto) {
@@ -50,6 +53,16 @@ export class RoleService {
   }
 
   async deleteRoleById(id: string[]) {
+    const user = await this.userRepository
+      .createQueryBuilder('user')
+      .leftJoinAndSelect('user.roles', 'role')
+      .where('role.id in (:...ids)', { ids: id })
+      .getMany();
+    if (user.length) {
+      throw `${user
+        .map((item) => item.username)
+        .join(',')}账户已绑定该菜单，无法删除`;
+    }
     return await this.roleRepository.delete(id);
   }
 }
